set up GitHub Actions
CI script will automatically build the program and run tests (if we add them) upon pull requests to any branch and upon push to main. In particular, this should complain whenever cargo does while building the result of the merge, and the result should show up on the PR page.